backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.0
pandas==2.1.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
pyarrow==13.0.0
